CODE RUB: Move Lab Services Under Labs Namespace
