+++ DMX.Core.Api.Tests.Unit/Services/Foundations/LabServiceTests.cs
@@ -11,7 +11,7 @@ using DMX.Core.Api.Brokers.Loggings;
 using DMX.Core.Api.Brokers.ReverbApis;
 using DMX.Core.Api.Models.Externals.ExternalLabs;
 using DMX.Core.Api.Models.Labs;
-using DMX.Core.Api.Services.Foundations;
+using DMX.Core.Api.Services.Foundations.Labs;
 using KellermanSoftware.CompareNetObjects;
 using Moq;
 using RESTFulSense.Exceptions;
