+++ DMX.Core.Api/Services/Foundations/Labs/ILabService.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using DMX.Core.Api.Models.Labs;
 
-namespace DMX.Core.Api.Services.Foundations
+namespace DMX.Core.Api.Services.Foundations.Labs
 {
     public interface ILabService
     {
