+++ DMX.Core.Api/Startup.cs
@@ -6,6 +6,7 @@ using System.Text.Json.Serialization;
 using DMX.Core.Api.Brokers.Loggings;
 using DMX.Core.Api.Brokers.ReverbApis;
 using DMX.Core.Api.Services.Foundations;
+using DMX.Core.Api.Services.Foundations.Labs;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
@@ -29,11 +30,11 @@ namespace DMX.Core.Api
             services.AddHttpClient();
             services.AddLogging();
             AddBrokers(services);
-            services.AddTransient<ILabService, LabService>();
+            AddServices(services);
 
-            services.AddSwaggerGen(c =>
+            services.AddSwaggerGen(option =>
             {
-                c.SwaggerDoc(
+                option.SwaggerDoc(
                     name: "v1",
                     info: new OpenApiInfo
                     {
@@ -69,5 +70,8 @@ namespace DMX.Core.Api
             services.AddTransient<ILoggingBroker, LoggingBroker>();
             services.AddTransient<IReverbApiBroker, ReverbApiBroker>();
         }
+
+        private static void AddServices(IServiceCollection services) =>
+            services.AddTransient<ILabService, LabService>();
     }
 }
